+++ src/models/Settings.php
@@ -38,6 +38,21 @@ class Settings extends Model
     public $volumeUid;
 
     /**
+     * @var bool
+     */
+    public $enableProofs = true;
+
+    /**
+     * @var bool
+     */
+    public $showAllLineItemsOnOrderTab = true;
+
+    /**
+     * @var bool
+     */
+    public $showOptionsOnOrderTab = true;
+
+    /**
      * @var string
      */
     public $volumeSubpath;
@@ -71,7 +86,16 @@ class Settings extends Model
      */
     public function rules()
     {
+
         return [
+            ['volumeUid', 'required'],
+            [
+                [
+                    'enableProofs',
+                    'showOptionsOnOrderTab',
+                ],
+                'boolean'
+            ],
             [
                 [
                     'volumeUid',
@@ -82,16 +106,6 @@ class Settings extends Model
                     'proofEmailUid',
                 ],
                 'string'
-            ],
-            [
-                [
-                    'volumeUid',
-                    'proofsSentStatusUid',
-                    'proofsApprovedStatusUid',
-                    'proofsRejectedStatusUid',
-                    'proofEmailUid',
-                ],
-                'required'
             ],
         ];
     }